pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "colmena_swarm_pm"
dynamic = ["version"]
authors = [
    { name = "Workflows and Distributed Computing group" },
]
description = "COLMENA Programming model library"
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
]
homepage = "https://proyecto-colmena.com/en"
dependencies = [
    "coloredlogs",
    "typing",
    "eclipse-zenoh==1.0.1",
]

[project.optional-dependencies]
ci = [
    "pytest"
]
role = [
    "eclipse-zenoh==1.0.1",
    "colmena-zeromq-pyre==0.3.5",
    "numpy==2.0.2"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "colmena/__init__.py"

[project.scripts]
colmena_build = "colmena.building_tool.colmena_build:main"

[tool.hatch.build]
include = [
    "colmena/*"
]